requirements.txt
torch>=1.10.0
torchvision
streamlit
Pillow
